fix(tests): split stale boss-model coverage from live supervisor tests
